### 2023-01-06 #5047 @editorialbot
[PRE REVIEW]: Fast Resampling and Monte Carlo Methods in Python
**Submitting author:** @mdhaber (<a href="http://orcid.org/0000-0003-4806-3601">Matt Haberland</a>)
**Repository:** https://github.com/mdhaber/scipy
**Branch with paper.md** (empty if default branch): joss_resampling
**Version:** 1.11.0.dev0+1225.ca8d481
