### 2023-01-09 c2643a2
continuing work on chandrupatla, adjusting flags and status output message
