### 2023-01-10 #5047 @gkthiruvathukal
@jbytecode While awaiting your answers, we're going to initiate a query scope here.
