### 2023-01-10 #5047 @mdhaber
Thank you for your questions, @gkthiruvathukal. To begin, perhaps I should include the "Additional comments" I entered with my submission:

> The URL submitted here is to a branch of my fork of the SciPy library. Since the contributions are parts of this much larger library, the following links may be helpful when assessing the magnitude of the contribution:
> * Monte Carlo Test PRs: https://github.com/scipy/scipy/pulls?> q=is%3Apr+label%3Ascipy.stats+is%3Amerged+author%3Amdhaber+monte_carlo_test
> * Permutation Test PRs: https://github.com/scipy/scipy/pulls?q=is%3Apr+label%3Ascipy.stats+is%3Amerged+author%3Amdhaber+permutation_test
> * Bootstrap PRs: https://github.com/scipy/scipy/pulls?q=is%3Apr+label%3Ascipy.stats+is%3Amerged+author%3Amdhaber+bootstrap
> * Implementation Code: https://github.com/scipy/scipy/blob/main/scipy/stats/_resampling.py
> * Test Suite: https://github.com/scipy/scipy/blob/main/scipy/stats/tests/test_resampling.py

Since there are several questions about my fork of SciPy, it may also help to clarify that this manuscript is not about my fork of SciPy or another a new repository. The manuscript claims the enhancements (listed above) to the larger SciPy library.

This is based on [Should I write my own software or contribute to an existing package?](https://joss.readthedocs.io/en/latest/submitting.html#should-i-write-my-own-software-or-contribute-to-an-existing-package).

![image](https://user-images.githubusercontent.com/6570539/211627308-3389aeab-8587-448e-80f5-aa112bca6ee3.png)

---

Responses to individual questions:

> The repo is forked from the scipy/scipy repo and the forked one seems to be quite young.
> It seems you are one of the prominent contributors of the scipy main repo.
> In what extend you forked the repo with a few new contributions?

The features claimed in the JOSS manuscript have been under development for about two years, starting with scipy/scipy#13371.

The `joss_resampling` branch in my fork of the SciPy repository is identical to SciPy 1.10.0, released last week, except for the addition of the JOSS manuscript. I took this approach based on "[Submitting a paper to JOSS](https://joss.readthedocs.io/en/latest/submitting.html)".

![image](https://user-images.githubusercontent.com/6570539/211626828-c5701149-f675-45bd-94fe-66ff26b447a8.png)

The `joss_resampling` branch is temporary; it is only for the submission of this manuscript.

> Does the forked repo improve the previous one?
> What is the point of creating a fork of a repo you are already a prominent contributor?

I use my fork of SciPy, [mdhaber/scipy,](https://github.com/mdhaber/scipy) to submit pull requests to the main repository, [scipy/scipy](https://github.com/scipy/scipy). The manuscript is about features added by pull requests to the SciPy main repository; all of which are available in SciPy 1.10.0; see e.g. [`scipy.stats.bootstrap`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.bootstrap.html), [`scipy.stats.permutation_test`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.permutation_test.html#scipy.stats.permutation_test), and [`scipy.stats.monte_carlo_test`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.monte_carlo_test.html#scipy.stats.monte_carlo_test).

> The manuscript is written in a way that the scipy is only for resampling methods. Why? Is SciPy created for resampling?

SciPy is a very broad scientific computing package for Python. These resampling methods were added to the `scipy.stats` subpackage of SciPy. To avoid this confusion, I included statements like:

![image](https://user-images.githubusercontent.com/6570539/211626119-46dbca51-cb67-4541-b468-c3f29b0f82ae.png)

Please let me know how I can clarify this in the manuscript.

> Why haven't you submit a scipy manuscript with other contributors?

I am the only author of the contributions claimed by the manuscript. In the Acknowledgements section of the manuscript, I thanked those who reviewed my pull requests. The other features of SciPy, which have other authors, are not claimed by the manuscript.
