### 2023-01-11 #5047 @gkthiruvathukal
@mdhaber I am actually satisfied with your responses. I'm going to ask @arfon to chime in here, because I don't have experience as EIC or editor with this type of submission. I think for me the biggest question is whether this "fork" is likely to be contributed upstream, perhaps as a pull request. This is a concern I have for any substantial fork, not just your project. If not, is there some plan to package it so it "tracks" SciPy closely enough so it can remain useful as a standalone package (or plugin, if applicable).  I am worried that few researchers would use it in its current form (even if it is awesome otherwise) if is not well-integrated within the ecosystem.

Thanks @jbytecode for having me chime in. I think it is mostly compliant with our criteria, so it is a matter of getting a second look from Arfon before we proceed to edit it.
