### 2023-01-12 #5047 @arfon
Thanks for the input @mdhaber and ping @gkthiruvathukal and @jbytecode. So just to clarify, the contributions outlined in this submission have already made it 'upstream' back to `scipy/scipy`?
