### 2023-01-16 cc7dd19
MAINT: stats.pearsonr: correct p-value with negative correlation
