### 2023-01-16 #5047 @arfon
@gkthiruvathukal – I think this is OK to proceed. @gkthiruvathukal – One thing we're going to need to do here is make it easy for reviewers to understand _what_ exactly we're asking them to review.

@mdhaber – your help here outlining the exact capabilities added to SciPy is appreciated :-)
