### 2023-01-16 #5047 @gkthiruvathukal
@arfon I agree it is important for them to know what they are reviewing. I believe this is already covered in @mdhaber's response to my query above. I tagged it with the 🎉. Reviewers can see the items of focus beginning with the "I use my fork of SciPy" sentence, which lists the claimed features associated with this JOSS submission.

@jbytecode Assuming your reviewers are available, I think you can remove the waitlist tag.
