### 2023-01-18 96de5d5
ENH: scipy._lib._util: add Parallelizer
