### 2023-01-18 9039898
Fixed more pep8 violations
