### 2023-01-18 #5047 @jbytecode
👋👋👋 Dear @coatless and @davismcc 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Fast Resampling and Monte Carlo Methods in Python**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5047).

Thank you in advance!
