### 2023-01-19 a6bda04
TST: stats: add generic entropy accuracy test
