### 2023-01-20 #5047 @jbytecode
👋👋👋 Dear @kose-y 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Fast Resampling and Monte Carlo Methods in Python**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5047).

We've already assigned the first reviewer. We need a second one. I hope you accept our invitation.

Thank you in advance!
