### 2023-01-22 #5092 @editorialbot
[REVIEW]: Fast Resampling and Monte Carlo Methods in Python
**Submitting author:** @mdhaber (<a href="http://orcid.org/0000-0003-4806-3601">Matt Haberland</a>)
**Repository:** https://github.com/mdhaber/scipy
**Branch with paper.md** (empty if default branch): joss_resampling
**Version:** v1.10.0-joss-article
