### 2023-01-24 890d1cd
STY/DOC: stats.dirichlet_multinomial: PEP8 and doc fixes
