### 2023-01-24 bd5d9f8
STY: stats.dirichlet_multinomial: more PEP8
