### 2023-01-24 de26dfa
MAINT: stats.dirichlet_multinomial: one more PEP8
