+++ scipy/stats/_multivariate.py
@@ -5894,7 +5894,7 @@ class dirichlet_multinomial_gen(multi_rv_generic):
         if np.less(x, 0).any():
             raise ValueError("`x` must not contain a non-negative integer.")
 
-        # Checks to make sure that every number in x can be expressed as an int.
+        # Checks to make sure that every number in x can be expressed as an int
         if not np.equal(np.mod(x, 1), 0).all():
             raise ValueError("`x` must only contain integers.")
 
