### 2023-01-24 5fc934b
MAINT: stats.dirichlet_multinomial: adjustments
