+++ scipy/stats/_multivariate.py
@@ -1387,14 +1387,6 @@ def _dirichlet_check_parameters(alpha):
     return alpha
 
 
-def _dirichlet_multinomial_check_parameters(alpha):
-    alpha = np.asarray(alpha)
-    if np.min(alpha) <= 0:
-        raise ValueError("All parameters must be greater than 0")
-
-    return alpha
-
-
 def _dirichlet_check_input(alpha, x):
     x = np.asarray(x)
 
@@ -5751,8 +5743,66 @@ def _sample_uniform_direction(dim, size, random_state):
     return samples
 
 
+_dirichlet_mn_doc_default_callparams = """\
+alpha : array_like
+    The concentration parameters. The number of entries determines the
+    dimensionality of the distribution.
+n : array_like
+    The number of trials.
+"""
+
+_dirichlet_mn_doc_frozen_callparams = ""
+
+_dirichlet_mn_doc_frozen_callparams_note = """\
+See class definition for a detailed description of parameters."""
+
+dirichlet_mn_docdict_params = {
+    '_dirichlet_mn_doc_default_callparams': _dirichlet_mn_doc_default_callparams,  # noqa
+    '_doc_random_state': _doc_random_state
+}
+
+dirichlet_mn_docdict_noparams = {
+    '_dirichlet_mn_doc_default_callparams': _dirichlet_mn_doc_frozen_callparams,
+    '_doc_random_state': _doc_random_state
+}
+
+def _dirichlet_multinomial_check_parameters(alpha, n, x=None):
+
+    if x is None:
+        alpha = np.asarray(alpha)
+    else:
+        # Ensure that `x` and `alpha` are arrays. If the shapes are
+        # incompatible, NumPy will raise an appropriate error.
+        x, alpha = np.broadcast_arrays(x, alpha)
+
+        x_int = np.floor(x)
+        if np.any(x <= 0) or np.any(x != x_int):
+            raise ValueError("`x` must contain only non-negative integers.")
+        x = x_int
+
+    alpha = np.asarray(alpha)
+
+    if np.any(alpha < 0):
+        raise ValueError("`alpha` must contain only positive values.")
+
+    n_int = np.floor(n)
+    if np.any(n <= 0) or np.any(n != n_int):
+        raise ValueError("`n` must be non-negative integer(s).")
+    n = n_int
+
+    sum_alpha = np.sum(alpha, axis=-1)
+    sum_alpha, n = np.broadcast_arrays(sum_alpha, n)
+
+    return (alpha, sum_alpha, n) if x is None else (alpha, sum_alpha, n, x)
+
+
 class dirichlet_multinomial_gen(multi_rv_generic):
     r"""A Dirichlet multinomial random variable.
+
+    The Dirichlet multinomial distribution is a compound probability
+    distribution: it is the multinomial distribution with number of trials
+    `n` and class probabilities ``p`` randomly sampled from a Dirichlet
+    distribution with concentration parameters `alpha`.
 
     Methods
     -------
@@ -5769,7 +5819,7 @@ class dirichlet_multinomial_gen(multi_rv_generic):
 
     Parameters
     ----------
-    %(_dirichlet_doc_default_callparams)s
+    %(_dirichlet_mn_doc_default_callparams)s
     %(_doc_random_state)s
 
     See Also
@@ -5794,37 +5844,35 @@ class dirichlet_multinomial_gen(multi_rv_generic):
 
     Get the PMF
 
-    >>> import numpy as np
-    >>> alpha = np.array([3, 4, 5])
-    >>> x = np.array([1, 2, 3])
-    >>> dirichlet_multinomial.pmf(alpha, x)
+    >>> n = 6  # number of trials
+    >>> alpha = [3, 4, 5]  # concentration parameters
+    >>> x = [1, 2, 3]  # counts
[…]
