### 2023-01-24 adf8254
MAINT: stats.dirichlet_multinomial: fix tests, remove broadcasting
