+++ scipy/stats/tests/test_multivariate.py
@@ -2917,18 +2917,14 @@ class TestDirichletMultinomial:
 
         alpha = rng.uniform(1e-10, 100, 10)
         x = rng.integers(0, 10, 10)
+        n = np.sum(x, -1)
 
-        n = np.array([np.sum(x, -1)])
-
-        d = dirichlet_multinomial(alpha)
-
-        assert_equal(d.logpmf(x), dirichlet_multinomial.logpmf(alpha, x))
-
-        assert_equal(d.pmf(x), dirichlet_multinomial.pmf(alpha, x))
-
-        assert_equal(d.mean(n), dirichlet_multinomial.mean(alpha, n))
-        assert_equal(d.var(n), dirichlet_multinomial.var(alpha, n))
-        assert_equal(d.cov(n), dirichlet_multinomial.cov(alpha, n))
+        d = dirichlet_multinomial(alpha, n)
+        assert_equal(d.logpmf(x), dirichlet_multinomial.logpmf(x, alpha, n))
+        assert_equal(d.pmf(x), dirichlet_multinomial.pmf(x, alpha, n))
+        assert_equal(d.mean(), dirichlet_multinomial.mean(alpha, n))
+        assert_equal(d.var(), dirichlet_multinomial.var(alpha, n))
+        assert_equal(d.cov(), dirichlet_multinomial.cov(alpha, n))
 
     def test_covariance_diagonal(self):
         # Makes sure that the diagonal of the
@@ -2962,127 +2958,55 @@ class TestDirichletMultinomial:
 
     def test_pmf(self):
         x = np.array([1, 2, 3])
+        n = np.sum(x)
         alpha = np.array([3, 4, 5])
         y = 0.08484162895927604
-        y_1 = dirichlet_multinomial.pmf(alpha, x)
+        y_1 = dirichlet_multinomial.pmf(x, alpha, n)
         assert_almost_equal(y, y_1)
 
     def test_logpmf(self):
         x = np.array([1, 2, 3])
+        n = np.sum(x)
         alpha = np.array([3, 4, 5])
-        y = dirichlet_multinomial.pmf(alpha, x)
-        y_1 = dirichlet_multinomial.logpmf(alpha, x)
+        y = dirichlet_multinomial.pmf(x, alpha, n)
+        y_1 = dirichlet_multinomial.logpmf(x, alpha, n)
         assert_equal(y, np.exp(y_1))
 
     def test_negative_x(self):
         x = np.array([1, -1, 3])
         alpha = np.array([3, 4, 5])
-        text = "`x` must not contain a non-negative integer."
+        text = "`x` must contain only non-negative integers."
         with assert_raises(ValueError, match=text):
-            dirichlet_multinomial.logpmf(alpha, x)
+            dirichlet_multinomial.logpmf(x, alpha, x.sum())
 
     def test_float_x(self):
         x = np.array([1, 1.3, 3])
         alpha = np.array([3, 4, 5])
-        text = "`x` must only contain integers."
+        text = "`x` must contain only non-negative integers."
         with assert_raises(ValueError, match=text):
-            dirichlet_multinomial.logpmf(alpha, x)
+            dirichlet_multinomial.logpmf(x, alpha, x.sum())
 
     def test_alpha_with_zero(self):
         alpha = np.array([1, 0, 2])
         x = np.array([1, 2, 3])
-        text = "All parameters must be greater than 0"
+        text = "`alpha` must contain only positive values."
         with assert_raises(ValueError, match=text):
-            dirichlet_multinomial.logpmf(alpha, x)
+            dirichlet_multinomial.logpmf(x, alpha, x.sum())
 
[…]
