### 2023-01-26 07881cc
TST: stats.dirichlet_multinomial: strengthen tests
