+++ scipy/stats/tests/test_multivariate.py
@@ -29,6 +29,7 @@ from scipy.stats import (multivariate_normal, multivariate_hypergeom,
                          random_table, uniform_direction,
                          dirichlet_multinomial)
 from scipy.stats import _covariance, Covariance
+from scipy import stats
 
 from scipy.integrate import romb
 from scipy.special import multigammaln
@@ -2912,12 +2913,20 @@ def test_random_state_property():
 
 
 class TestDirichletMultinomial:
-    def test_frozen(self):
-        rng = np.random.default_rng(2846)
+    @classmethod
+    def get_params(self, m):
+        rng = np.random.default_rng(28469824356873456)
+        alpha = rng.uniform(0, 100, size=2)
+        x = rng.integers(1, 10, size=(m, 2))
+        n = x.sum(axis=-1)
+        return rng, m, alpha, n, x
 
-        alpha = rng.uniform(1e-10, 100, 10)
+    def test_frozen(self):
+        rng = np.random.default_rng(28469824356873456)
+
+        alpha = rng.uniform(0, 100, 10)
         x = rng.integers(0, 10, 10)
-        n = np.sum(x, -1)
+        n = np.sum(x, axis=-1)
 
         d = dirichlet_multinomial(alpha, n)
         assert_equal(d.logpmf(x), dirichlet_multinomial.logpmf(x, alpha, n))
@@ -2926,92 +2935,102 @@ class TestDirichletMultinomial:
         assert_equal(d.var(), dirichlet_multinomial.var(alpha, n))
         assert_equal(d.cov(), dirichlet_multinomial.cov(alpha, n))
 
-    def test_covariance_diagonal(self):
-        # Makes sure that the diagonal of the
-        # covariance matrix is the variance.
-        alpha = [1, 5, 3]
-        n = np.array([3])
-        c = dirichlet_multinomial.cov(alpha, n)
-        v = dirichlet_multinomial.var(alpha, n)
-
-        assert_array_equal(v, c.diagonal())
-
-    def test_variance(self):
-        alpha = [0.3, 2.1, 5.4]
-        n = 3
-        # Truncated to the 6th digit, so not
-        # exactly what we would expect from var
-        expected_v = [0.136162, 0.724381, 0.784293]
-        v = dirichlet_multinomial.var(alpha, n)
-        assert_array_almost_equal(expected_v, v)
-
-    def test_covariance(self):
-        alpha = [0.3, 2.1, 5.4]
-        n = np.array([3])
-        # Truncated to the 6th digit, so not
-        # exactly what we would expect from var
-        expected_c = np.array([[0.136162, -0.038125, -0.098037],
-                               [-0.038125, 0.724381, -0.686256],
-                               [-0.098037, -0.686256, 0.784293]])
-        c = dirichlet_multinomial.cov(alpha, n)
-        assert_array_almost_equal(expected_c, c)
-
-    def test_pmf(self):
+    def test_pmf_logpmf_against_R(self):
+        # # Compare PMF against R's extraDistr ddirmnon
+        # # library(extraDistr)
+        # # options(digits=16)
+        # ddirmnom(c(1, 2, 3), 6, c(3, 4, 5))
         x = np.array([1, 2, 3])
         n = np.sum(x)
         alpha = np.array([3, 4, 5])
-        y = 0.08484162895927604
-        y_1 = dirichlet_multinomial.pmf(x, alpha, n)
-        assert_almost_equal(y, y_1)
+        res = dirichlet_multinomial.pmf(x, alpha, n)
+        logres = dirichlet_multinomial.logpmf(x, alpha, n)
+        ref = 0.08484162895927638
+        assert_allclose(res, ref)
[…]
