### 2023-01-26 c9be90a
MAINT: stats.dirichlet_multinomial: edge case handling
