### 2023-01-26 42b4589
DOC: Added missing keyword parameters for dirchlet_multinomial
