### 2023-01-27 5a54be3
TST: stats.dirichlet_multinomial: fix tests
