### 2023-01-27 071eb7c
STY: stats.dirichlet_multinomial: fix lint
