### 2023-01-30 b2045b5
MAINT: stats: convert boost overflow warnings to errors
