+++ scipy/stats/_discrete_distns.py
@@ -354,11 +354,11 @@ class nbinom_gen(rv_discrete):
         return _boost._nbinom_sf(k, n, p)
 
     def _isf(self, x, n, p):
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _boost._nbinom_isf(x, n, p)
 
     def _ppf(self, q, n, p):
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _boost._nbinom_ppf(q, n, p)
 
     def _stats(self, n, p):
@@ -1418,7 +1418,7 @@ class skellam_gen(rv_discrete):
                 random_state.poisson(mu2, n))
 
     def _pmf(self, x, mu1, mu2):
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             px = np.where(x < 0,
                           _boost._ncx2_pdf(2*mu2, 2*(1-x), 2*mu1)*2,
                           _boost._ncx2_pdf(2*mu1, 2*(1+x), 2*mu2)*2)
@@ -1427,7 +1427,7 @@ class skellam_gen(rv_discrete):
 
     def _cdf(self, x, mu1, mu2):
         x = floor(x)
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             px = np.where(x < 0,
                           _boost._ncx2_cdf(2*mu2, -2*x, 2*mu1),
                           1 - _boost._ncx2_cdf(2*mu1, 2*(x+1), 2*mu2))
