+++ scipy/stats/_continuous_distns.py
@@ -689,13 +689,11 @@ class beta_gen(rv_continuous):
         return _boost._beta_sf(x, a, b)
 
     def _isf(self, x, a, b):
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _boost._beta_isf(x, a, b)
 
     def _ppf(self, q, a, b):
-        with np.errstate(over='ignore'):  # see gh-17432
-            message = "overflow encountered in _beta_ppf"
-            warnings.filterwarnings('ignore', message=message)
+        with np.errstate(over='raise'):  # see gh-17432
             return _boost._beta_ppf(q, a, b)
 
     def _stats(self, a, b):
@@ -6653,31 +6651,31 @@ class ncx2_gen(rv_continuous):
 
     def _pdf(self, x, df, nc):
         cond = np.ones_like(x, dtype=bool) & (nc != 0)
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _lazywhere(cond, (x, df, nc), f=_boost._ncx2_pdf,
                               f2=lambda x, df, _: chi2._pdf(x, df))
 
     def _cdf(self, x, df, nc):
         cond = np.ones_like(x, dtype=bool) & (nc != 0)
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _lazywhere(cond, (x, df, nc), f=_boost._ncx2_cdf,
                               f2=lambda x, df, _: chi2._cdf(x, df))
 
     def _ppf(self, q, df, nc):
         cond = np.ones_like(q, dtype=bool) & (nc != 0)
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _lazywhere(cond, (q, df, nc), f=_boost._ncx2_ppf,
                               f2=lambda x, df, _: chi2._ppf(x, df))
 
     def _sf(self, x, df, nc):
         cond = np.ones_like(x, dtype=bool) & (nc != 0)
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _lazywhere(cond, (x, df, nc), f=_boost._ncx2_sf,
                               f2=lambda x, df, _: chi2._sf(x, df))
 
     def _isf(self, x, df, nc):
         cond = np.ones_like(x, dtype=bool) & (nc != 0)
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _lazywhere(cond, (x, df, nc), f=_boost._ncx2_isf,
                               f2=lambda x, df, _: chi2._isf(x, df))
 
@@ -6759,14 +6757,14 @@ class ncf_gen(rv_continuous):
         return _boost._ncf_cdf(x, dfn, dfd, nc)
 
     def _ppf(self, q, dfn, dfd, nc):
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _boost._ncf_ppf(q, dfn, dfd, nc)
 
     def _sf(self, x, dfn, dfd, nc):
         return _boost._ncf_sf(x, dfn, dfd, nc)
 
     def _isf(self, x, dfn, dfd, nc):
-        with np.errstate(over='ignore'):  # see gh-17432
+        with np.errstate(over='raise'):  # see gh-17432
             return _boost._ncf_isf(x, dfn, dfd, nc)
 
     def _munp(self, n, dfn, dfd, nc):
[…]
