### 2023-02-03 e2690ca
TST: stats.betaprime: simplify tests
