### 2023-02-11 473475c
TST: stats: add tests for other isses resolved by gh-17432
