### 2023-02-14 761b9c5
Updating _chandrupatla in _zeros_py.py with review feedback, closes gh #7242
