+++ scipy/optimize/_zeros_py.py
@@ -11,7 +11,7 @@ _xtol = 2e-12
 _rtol = 4 * np.finfo(float).eps
 
 __all__ = ['newton', 'bisect', 'ridder', 'brentq', 'brenth', 'toms748',
-           'chandrupatla', 'RootResults']
+           '_chandrupatla', 'RootResults']
 
 # Must agree with CONVERGED, SIGNERR, CONVERR, ...  in zeros.h
 _ECONVERGED = 0
@@ -1396,51 +1396,85 @@ def toms748(f, a, b, args=(), k=1,
     return _results_select(full_output, (x, function_calls, iterations, flag))
 
 
-def chandrupatla(f, x0, x1, args=(), xtol=_xtol,
-                 rtol=_rtol, maxiter=_iter, callback=None):
-
+def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
+                 xatol=_rtol, maxiter=_iter, callback=None):
     """
-    As written in [Sachs2015]_ which in turn is based on Chandrupatla's
-    algorithm as described in [Scherer2018]_.
-    This allows vector arguments for x0, x1, and args
+    Find the root of an elementwise function using Chandrupatla's algorithm.
+
+    This function allows for `x0`, `x1`, amd the output of `f` to be of any
+    broadcastable shapes. For each element of the output of `f`, `chandrupatla`
+    seeks the scalar root that makes the element 0.
 
     Parameters
     ----------
 
     f : callable
-        The function whose zero is wanted. It must be a function of a
-        single variable of the form ``f(x,a,b,c...)``, where ``a,b,c...``
-        are extra arguments that can be passed in the `args` parameter.
+        The elementwise function whose root is wanted. The signature must be::
+
+            f(x: ndarray, *args) -> ndarray
+
+         where each element of ``x`` is a finite real and ``args`` is a tuple
+         with an arbitrary number of components of any type(s).
+         `chandrupatla` seeks an array ``x`` (or scalar) such that each
+         corresponding element in the output array (or scalar) are zero.
     x0 : float, sequence, or ndarray
-        An initial estimate of the zero that should be somewhere near the
-        actual zero. If not scalar, then `func` must be vectorized and return
-        a sequence or array of the same shape as its first argument.
-    x1 : float, optional
-        Another estimate of the zero that should be somewhere near the
-        actual zero.
-    xtol : float, optional
-        Tolerance (absolute) for termination.
-    rtol : float, optional
-        Tolerance (relative) for termination.
+        The lower bound of the root of the function. Follows normal
+        broadcasting rules and will be broadcast against `x1`and `f` before
+        being passed into `f`.
+    x1 : float, sequence, or ndarray
+        The upper bound of the root of the function. Follows normal
+        broadcasting rules and will be broadcast against `x0` and `f` before
+        being passed into `f`.
+    xatol : float, optional
+        Tolerance (absolute) for termination. Termination occurs if
+        `2*xatol*np.abs(r) + xrtol / np.abs(b - c) > 0.5`.
+    xrtol : float, optional
+        Tolerance (relative) for termination. Termination occurs if
+        `2*xatol*np.abs(r) + xrtol / np.abs(b - c) > 0.5`.
     maxiter : int, optional
-        If convergence is not achieved in `maxiter` iterations, an error is
-        raised. Must be >= 0.
+        If convergence is not achieved in `maxiter` iterations, the corresponding
+        elements of the `status` and `success` attributes will indicate the iteration
+        limit was reached.
[…]

[…]
