### 2023-02-25 a32db7d
changing output to RootResults object
