### 2023-02-26 #5092 @mdhaber
Since the paper is about specific contributions to a much larger library (as discussed in https://github.com/openjournals/joss-reviews/issues/5047#issuecomment-1377678757), I thought it might be helpful to point out some relevant files.

**License**: SciPy's license file is available at https://github.com/scipy/scipy/blob/main/LICENSE.txt.

**Contributions**: The "blame" of [`scipy.stats._resampling.py`](https://github.com/scipy/scipy/blame/main/scipy/stats/_resampling.py) and [`scipy.stats.tests.test_resampling.py`](https://github.com/scipy/scipy/blame/main/scipy/stats/tests/test_resampling.py) reveal some of my contributions to these features over the past two years.

**Installation**: Installation instructions for the whole library are at https://scipy.org/install/. The list of dependencies is available in the [environment.yml ](https://github.com/scipy/scipy/blob/main/environment.yml) file.

**Documentation**:
The [scipy.stats API reference](https://docs.scipy.org/doc/scipy/reference/stats.html#resampling-methods) includes both input/output documentation and examples. More realistic usage examples are available in the [scipy.stats tutorials](https://docs.scipy.org/doc/scipy/tutorial/stats/resampling.html).

**Tests**: Automated tests are in [`scipy.stats.tests.test_resampling.py`](https://github.com/scipy/scipy/blob/main/scipy/stats/tests/test_resampling.py). Continuous integration test results can be seen in any [pull request to SciPy](https://github.com/scipy/scipy/pulls), or tests can be run locally using `import scipy.stats; scipy.stats.test(verbose=2)`.

**Community Guidlines**: Community guidelines are available in the [Developer Documentation](https://scipy.github.io/devdocs/dev/conduct/code_of_conduct.html).

I hope this helps!

@jbytecode would it be appropriate for me to also suggest how the functionality and performance claims could be verified?
