### 2023-02-26 4892d8b
modifying RootResult to accept a None flag
