### 2023-02-27 #5092 @mdhaber
Thanks @jbytecode I added notes on how one might begin to verify the functionality and performance claims. I'd be happy to provide other examples if it would help!
