### 2023-03-08 a5f707a
MAINT: stats.goodness_of_fit: draft Shapiro-Wilk
