### 2023-03-09 27b4e87
MAINT: stats.goodness_of_fit: begin to correct order statistic variance
