### 2023-03-10 d9f2817
MAINT: stats.goodness_of_fit: fix shapiro-wilk statistic
