### 2023-03-10 8c09984
MAINT: add test against generic fit method for vonmises
