### 2023-03-10 313b18f
TST: stats.vonmises: rename scale-> shape, simplify, mark xslow
[skip ci]
