### 2023-03-10 309edd2
TST: stats.vonmises.fit: correct test
[skip ci]
