+++ scipy/stats/tests/test_distributions.py
@@ -213,7 +213,7 @@ class TestVonMises:
             kwds['floc'] = rvs_loc
 
         _assert_less_or_close_loglike(stats.vonmises, data,
-                                      stats.vonmises.nnlf, **kwds)
+                                      stats.vonmises.nnlf, 1e-6, **kwds)
 
     @pytest.mark.parametrize('loc', [-0.5 * np.pi, 0, np.pi])
     @pytest.mark.parametrize('kappa', [1, 10, 100, 1000])
@@ -252,7 +252,7 @@ class TestVonMises:
         assert -np.pi < loc_fit < np.pi
 
 
-def _assert_less_or_close_loglike(dist, data, func, **kwds):
+def _assert_less_or_close_loglike(dist, data, func, rtol=1e-15, **kwds):
     """
     This utility function checks that the log-likelihood (computed by
     func) of the result computed using dist.fit() is less than or equal
@@ -262,10 +262,11 @@ def _assert_less_or_close_loglike(dist, data, func, **kwds):
     """
     mle_analytical = dist.fit(data, **kwds)
     numerical_opt = super(type(dist), dist).fit(data, **kwds)
+    #print(f"analytical: {mle_analytical}, super: {numerical_opt}")
     ll_mle_analytical = func(mle_analytical, data)
     ll_numerical_opt = func(numerical_opt, data)
     assert (ll_mle_analytical <= ll_numerical_opt or
-            np.allclose(ll_mle_analytical, ll_numerical_opt, rtol=1e-15))
+            np.allclose(ll_mle_analytical, ll_numerical_opt))
 
 
 def assert_fit_warnings(dist):
