### 2023-03-11 fa127c7
TST: only adjust tolerance for fixed loc parameter
