### 2023-03-13 d72d36c
updating testing with chandrupatla class, _zeros_py to add two methods for testing and change iteration loop, and added status messages to optimize dictionary
