+++ scipy/optimize/_zeros_py.py
@@ -2,7 +2,7 @@ import warnings
 from collections import namedtuple
 import operator
 from . import _zeros
-from ._optimize import OptimizeResult, _call_callback_maybe_halt
+from ._optimize import OptimizeResult, _call_callback_maybe_halt, _status_message
 import numpy as np
 
 
@@ -1391,7 +1391,8 @@ def toms748(f, a, b, args=(), k=1,
 
 
 def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
-                  xatol=_rtol, maxiter=_iter, callback=None):
+                  xatol=_rtol, maxiter=_iter, ensure_scalar_x=False,
+                  callback=None, test_switch=False):
     """Find the root of an elementwise function using Chandrupatla's algorithm.
 
     This function allows for `x0`, `x1`, amd the output of `f` to be of any
@@ -1418,6 +1419,8 @@ def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
         The upper bound of the root of the function. Follows normal
         broadcasting rules and will be broadcast against `x0` and `f` before
         being passed into `f`.
+    args : tuple, optional
+        Determines shape of output
     xatol : float, optional
         Tolerance (absolute) for termination. Termination occurs if
         `2*xatol*np.abs(r) + xrtol / np.abs(b - c) > 0.5`.
@@ -1527,11 +1530,37 @@ def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
     # which in turn is based on Chandrupatla's algorithm as described in
     # Scherer https://books.google.com/books?id=cC-8BAAAQBAJ&pg=PA95
 
+    # FOR TESTING ONLY
+    if ensure_scalar_x:
+        def f_wrapped(x, *args, **kwargs):
+            return np.asarray([f(x[0], *args, **kwargs)])
+    else:
+        f_wrapped = f
+
+    # FOR TESTING ONLY
+    def termination_function1():
+        """chandrupatla termination per paper"""
+        term = np.logical_or(terminate,
+                             np.logical_or(fm == 0, tlim > 0.5))
+        return term
+
+    def termination_function2():
+        """termination like brent / bisect for testing"""
+        delta = (xatol + xrtol * abs(r))
+        term = np.logical_or(terminate,
+                             np.logical_or(fm == 0,
+                                           abs(b - a) < delta))
+        return term
+
+    termination_function = (termination_function1
+                            if test_switch
+                            else termination_function2)
+
     # Initialization
     b = x0
     a = c = x1
-    fb = f(b, *args)
-    fa = fc = f(a, *args)
+    fb = f_wrapped(b, *args)
+    fa = fc = f_wrapped(a, *args)
     t = 0.5
     iterations = 0
 
@@ -1543,11 +1572,7 @@ def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
     iterations = np.zeros_like(a)
 
     # flag to check state of convergence
-    flag = np.full(shape, _EINPROGRESS)
-
-    # Initialize an array of False,
-    # determines whether we should do inverse quadratic interpolation
-    iqi = np.zeros(shape, dtype=bool)
+    flag = np.full(shape, "in_progress")
 
     funcalls = np.full(np.shape(flag), 0)
     terminate = False
@@ -1556,7 +1581,7 @@ def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
         # use t to linearly interpolate between a and b,
         # and evaluate this function as our newest estimate xt
         xt = a + t * (b - a)
-        ft = f(xt, *args)
[…]
