### 2023-03-17 8852f2f
MAINT: stats.vonmises.fit: generalize kappa likelihood equation for arbitrary location
