### 2023-03-17 309e758
MAINT: stats.vonmises.fit: allow user to fix shape
