### 2023-03-31 2975006
MAINT: stats._axis_nan_policy: fix dtype of NaNs produced by decorator
