### 2023-04-04 6449a15
ENH: stats: improve dunnett allowance calculation
