### 2023-04-07 3e4ed06
switching out default termination condition
