### 2023-04-11 4ef87f7
ENH: refactor plot of CI
