### 2023-04-13 #5092 @mdhaber
Thanks @SaranjeetKaur; I've attempted to address your comments in mdhaber/scipy#101.
