+++ paper.bib
@@ -87,3 +87,50 @@
   journal = {GitHub Issue},
   url = {https://github.com/scipy/scipy/issues/12315}
 }
+
+@misc{ye2020,
+  author = {Leihua Ye},
+  title = {Bootstrap and Statistical Inference in Python},
+  year = {2020},
+  month = {12},
+  publisher = {Medium},
+  journal = {Towards Data Science},
+  url = {https://towardsdatascience.com/a06d098a8bfd}
+}
+
+@misc{saenger2019,
+  author = {Victor Saenger},
+  title = {How To Assess Statistical Significance in Your Data with Permutation Tests},
+  year = {2019},
+  month = {10},
+  publisher = {Medium},
+  journal = {Towards Data Science},
+  url = {https://towardsdatascience.com/8bb925b2113d}
+}
+
+
+@misc{omar2021,
+  author = {Hamid Omar},
+  title = {A Guide on How to Simulate and Visualize Hypothesis Tests Using Python Code},
+  year = {2021},
+  month = {06},
+  publisher = {Medium},
+  journal = {Analytics Vidhya},
+  url = {https://medium.com/analytics-vidhya/66861e0d0f74}
+}
+
+@misc{resample2020,
+  author = {Daniel Saxton and Hans Dembinski},
+  title = {resample 1.0.0},
+  year  = {2020},
+  month = {8},
+  url   = {https://github.com/scikit-hep/resample/},
+}
+
+@misc{permute2021,
+  author = {Jarrod Millman},
+  title = {permute 0.2},
+  year  = {2021},
+  month = {06},
+  url   = {https://github.com/statlab/permute},
+}
