### 2023-05-07 e6a9527
TST: optimize._chandrupatla: begin to add tests
