+++ scipy/optimize/tests/test_zeros.py
@@ -10,7 +10,7 @@ from numpy.testing import (assert_warns, assert_,
 import numpy as np
 from numpy import finfo, power, nan, isclose, sqrt, exp, sin, cos
 
-
+from scipy import stats
 from scipy.optimize import (_zeros_py as zeros, newton, root_scalar,
                             OptimizeResult)
 
@@ -207,72 +207,17 @@ class TestBracketMethods(TestScalarRootFinders):
 
 class TestChandrupatla(TestScalarRootFinders):
 
-    @staticmethod
-    def f3(x, a):
-        return a-x**2
+    def f(self, q, dist, p):
+        return dist.cdf(q) - p
 
-    def test_array_chandrupatla(self):
-        """test _chandrupatla with array, similar to newton method"""
-
-        k = np.arange(1, 8)
-        result = zeros._chandrupatla(self.f3, 0, 3, args=(k, ))
-        x_expected = ([1., 1.41421356, 1.73205081,
-                      2., 2.23606798, 2.44948974,
-                      2.64575131])
-        assert_allclose(result.x, x_expected)
-
-    def test_chandrupatla_newton_similar(self):
-        """test that _chandrupatla is giving a similar result to newton"""
-
-        k = np.arange(1, 8)
-
-        result = zeros._chandrupatla(self.f3, 0, 3, args=(k[0],))
-        expected = zeros.newton(self.f3, x0=0, x1=3, args=(k[0],))
-        assert_allclose(result.x, expected)
-
-    def test_chandrupatla_brent_similar(self):
-        """test than chandrupatla gives equal scalar result to brentq"""
-        def f4(x):
-            return (x**3 - 1)
-
-        result = zeros._chandrupatla(f4, 0, 3, maxiter=50)
-        expected = zeros.brentq(f4, 0, 3)
-
-        assert_allclose(result.x, expected)
-
-    def test_chandrupatla_np_vectorize(self):
-        """test that _chandrupatla handles input with np.vectorize"""
-
-        k = np.arange(1, 8)
-
-        res = zeros._chandrupatla(self.f3, 0, 3,
-                                  args=(k,),
-                                  maxiter=50)
-
-        vec_res = zeros._chandrupatla(np.vectorize(self.f3), 0, 3,
-                                      args=(k,),
-                                      maxiter=50)
-
-        assert_allclose(res.fun, vec_res.fun)
-        assert_allclose(res.x, vec_res.x)
-
-    def test_chandrupatla_broadcasting(self):
-        """test that _chandrupatla broadcasts inputs correctly"""
-
-        x0 = np.zeros((2, 1, 1))
-        x1 = np.full((1, 2, 1), 3)
-        k = np.array([[[1, 2]]])
-
-        res = zeros._chandrupatla(self.f3, x0=x0, x1=x1,
-                                  args=(k,),
-                                  maxiter=50)
-
-        expected = ([[[1., 1.41421356],
-                      [1., 1.41421356]],
-                     [[1., 1.41421356],
-                      [1., 1.41421356]]])
-
-        assert_allclose(res.x, expected)
+    @pytest.mark.parametrize('p', [0.6, np.linspace(0.05, 0.95, 10)])
+    def test_basic(self, p):
+        dist = stats.norm()
+        with np.errstate(invalid='ignore'):
+            res = zeros._chandrupatla(self.f, -5, 5, args=(dist, p))
+        ref = dist.ppf(p)
+        np.testing.assert_allclose(res.root, ref)
+        assert res.root.shape == ref.shape
 
 
 class TestNewton(TestScalarRootFinders):
