### 2023-05-07 b9c7f9f
MAINT: optimize._chandrupatla: refactor based on original paper
