+++ scipy/optimize/_zeros_py.py
@@ -1394,19 +1394,18 @@ def toms748(f, a, b, args=(), k=1,
     return _results_select(full_output, (x, function_calls, iterations, flag))
 
 
-def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
-                  xatol=_rtol, maxiter=_iter, ensure_scalar_x=False,
-                  callback=None, test_switch=False):
+def _chandrupatla(func, x1, x2, *, args=(), xrtol=_xtol, xatol=_rtol,
+                  maxiter=_iter, callback=None, test_switch=False):
     """Find the root of an elementwise function using Chandrupatla's algorithm.
 
-    This function allows for `x0`, `x1`, amd the output of `f` to be of any
+    This function allows for `x1`, `x2`, amd the output of `f` to be of any
     broadcastable shapes. For each element of the output of `f`, `chandrupatla`
     seeks the scalar root that makes the element 0.
 
     Parameters
     ----------
 
-    f : callable
+    func : callable
         The elementwise function whose root is wanted. The signature must be::
 
             f(x: ndarray, *args) -> ndarray
@@ -1415,11 +1414,11 @@ def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
          with an arbitrary number of components of any type(s).
          `chandrupatla` seeks an array ``x`` (or scalar) such that each
          corresponding element in the output array (or scalar) are zero.
-    x0 : float, sequence, or ndarray
+    x1 : float, sequence, or ndarray
         The lower bound of the root of the function. Follows normal
         broadcasting rules and will be broadcast against `x1`and `f` before
         being passed into `f`.
-    x1 : float, sequence, or ndarray
+    x2 : float, sequence, or ndarray
         The upper bound of the root of the function. Follows normal
         broadcasting rules and will be broadcast against `x0` and `f` before
         being passed into `f`.
@@ -1512,10 +1511,10 @@ def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
     >>> y.x
     2.6457513110645907
 
-    >>> x0 = np.zeros((2,1,1))
-    >>> x1 = np.full((1,2,1), 3)
+    >>> x1 = np.zeros((2,1,1))
+    >>> x2 = np.full((1,2,1), 3)
     >>> k = np.array([[[1, 2]]])
-    >>> res = optimize._chandrupatla(f,x0=x0,x1=x1,
+    >>> res = optimize._chandrupatla(f,x1=x1,x2=x2,
                               args=(k,),
                               maxiter=50)
     >>> res.x
@@ -1534,151 +1533,119 @@ def _chandrupatla(f, x0, x1, *, args=(), xrtol=_xtol,
     # which in turn is based on Chandrupatla's algorithm as described in
     # Scherer https://books.google.com/books?id=cC-8BAAAQBAJ&pg=PA95
 
-    # FOR TESTING ONLY
-    if ensure_scalar_x:
-        def f_wrapped(x, *args, **kwargs):
-            return np.asarray([f(x[0], *args, **kwargs)])
-    else:
-        f_wrapped = f
+    def check_convergence(x1, f1, x2, f2, flag, eps_r=1e-5, eps_a=1e-10):
+        # See Section 4
+        xmin, fmin = x2.copy(), f2.copy()
+        imin = np.abs(f1) < np.abs(f2)
+        xmin[imin], fmin[imin] = x1[imin], f1[imin]
 
-    # FOR TESTING ONLY
-    def termination_function1():
-        """chandrupatla termination per paper"""
-        term = np.logical_or(terminate,
-                             np.logical_or(fm == 0, tlim > 0.5))
[…]
