### 2023-05-09 0c5146c
MAINT: optimize._chandrupatla: improvements and tests
