+++ scipy/optimize/_zeros_py.py
@@ -18,6 +18,7 @@ _ECONVERGED = 0
 _ESIGNERR = -1
 _ECONVERR = -2
 _EVALUEERR = -3
+_ECALLBACK = -4
 _EINPROGRESS = 1
 
 CONVERGED = 'converged'
@@ -1394,11 +1395,11 @@ def toms748(f, a, b, args=(), k=1,
     return _results_select(full_output, (x, function_calls, iterations, flag))
 
 
-def _chandrupatla(func, x1, x2, *, args=(), xrtol=_xtol, xatol=_rtol,
+def _chandrupatla(func, a, b, *, args=(), xrtol=_xtol, xatol=_rtol,
                   maxiter=_iter, callback=None, test_switch=False):
     """Find the root of an elementwise function using Chandrupatla's algorithm.
 
-    This function allows for `x1`, `x2`, amd the output of `f` to be of any
+    This function allows for `a`, `b`, amd the output of `f` to be of any
     broadcastable shapes. For each element of the output of `f`, `chandrupatla`
     seeks the scalar root that makes the element 0.
 
@@ -1406,24 +1407,20 @@ def _chandrupatla(func, x1, x2, *, args=(), xrtol=_xtol, xatol=_rtol,
     ----------
 
     func : callable
-        The elementwise function whose root is wanted. The signature must be::
+        The elementwise function whose root is desired. The signature must be::
 
-            f(x: ndarray, *args) -> ndarray
+            func(x: ndarray, *args) -> ndarray
 
-         where each element of ``x`` is a finite real and ``args`` is a tuple
-         with an arbitrary number of components of any type(s).
-         `chandrupatla` seeks an array ``x`` (or scalar) such that each
-         corresponding element in the output array (or scalar) are zero.
-    x1 : float, sequence, or ndarray
-        The lower bound of the root of the function. Follows normal
-        broadcasting rules and will be broadcast against `x1`and `f` before
-        being passed into `f`.
-    x2 : float, sequence, or ndarray
-        The upper bound of the root of the function. Follows normal
-        broadcasting rules and will be broadcast against `x0` and `f` before
-        being passed into `f`.
+         where each element of ``x`` is a finite real and ``args`` is a tuple,
+         which may contain an arbitrary number of components of any type(s).
+         `_chandrupatla` seeks an array (or scalar) ``x`` such that each
+         corresponding element in the output array (or scalar) is zero.
+    a, b : array_like
+        The lower and upper bounds of the root of the function. Follows normal
+        broadcasting rules and will be broadcasted against one another before
+        being passed into `func`.
     args : tuple, optional
-        Determines shape of output
+        Additional positional arguments to be passed to `func`.
     xatol : float, optional
         Tolerance (absolute) for termination. Termination occurs if
         `2*xatol*np.abs(r) + xrtol / np.abs(b - c) > 0.5`.
@@ -1431,67 +1428,53 @@ def _chandrupatla(func, x1, x2, *, args=(), xrtol=_xtol, xatol=_rtol,
         Tolerance (relative) for termination. Termination occurs if
         `2*xatol*np.abs(r) + xrtol / np.abs(b - c) > 0.5`.
     maxiter : int, optional
-        If convergence is not achieved in `maxiter` iterations, the
-        corresponding elements of the `status` and `success` attributes will
-        indicate the iteration
-        limit was reached.
+        The maximum number of iterations of the algorithm to perform.
     callback : callable, optional
[…]
