+++ scipy/optimize/_tstutils.py
@@ -704,17 +704,17 @@ fun4.root = 2
 
 
 def fun5(x):
-    return x**9
+    return x**9.
 fun5.root = 0
 
 
 def fun6(x):
-    return x**19
+    return x**19.
 fun6.root = 0
 
 
 def fun7(x):
-    return 0 if abs(x) < 3.8e-4 else x*np.exp(-x**(-2))
+    return 0 if abs(x) < 3.8e-4 else x*np.exp(-x**(-2.))
 fun7.root = 0
 
 
@@ -753,14 +753,14 @@ _CHANDRUPATLA_TESTS = [
     [fun2, [1e-12, 1e12], fun2.root, 48],
     [fun3, [0, 5], fun3.root, 21],
     [fun3, [-10, 10], fun3.root, 23],
-    [fun3, [1e-4, 1e4], fun3.root, 36],
-    [fun3, [1e-6, 1e6], fun3.root, 45],
-    [fun3, [1e-10, 1e10], fun3.root, 55],
+    [fun3, [-1e4, 1e4], fun3.root, 36],
+    [fun3, [-1e6, 1e6], fun3.root, 45],
+    [fun3, [-1e10, 1e10], fun3.root, 55],
     [fun4, [0, 5], fun4.root, 21],
     [fun4, [-10, 10], fun4.root, 23],
-    [fun4, [1e-4, 1e4], fun4.root, 33],
-    [fun4, [1e-6, 1e6], fun4.root, 43],
-    [fun4, [1e-10, 1e10], fun4.root, 54],
+    [fun4, [-1e4, 1e4], fun4.root, 33],
+    [fun4, [-1e6, 1e6], fun4.root, 43],
+    [fun4, [-1e10, 1e10], fun4.root, 54],
     [fun5, [-1, 4], fun5.root, 21],
     [fun5, [-2, 5], fun5.root, 22],
     [fun5, [-1, 10], fun5.root, 23],
@@ -781,7 +781,7 @@ _CHANDRUPATLA_TESTS = [
     [fun8, [2e-4, 9], fun8.root, 11],
     [fun8, [2e-4, 27], fun8.root, 12],
     [fun8, [2e-4, 81], fun8.root, 14],
-    [fun9, [2e-4, 2], fun9.root, 7],
+    [fun9, [2e-4, 1], fun9.root, 7],
     [fun9, [2e-4, 3], fun9.root, 8],
     [fun9, [2e-4, 9], fun9.root, 10],
     [fun9, [2e-4, 27], fun9.root, 11],
