### 2023-05-09 b76670c
MAINT: optimize._chandrupatla: adjust termination tolerances
