+++ scipy/optimize/_zeros_py.py
@@ -1395,13 +1395,13 @@ def toms748(f, a, b, args=(), k=1,
     return _results_select(full_output, (x, function_calls, iterations, flag))
 
 
-def _chandrupatla(func, a, b, *, args=(), xrtol=_xtol, xatol=_rtol,
-                  maxiter=_iter, callback=None, test_switch=False):
+def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
+                  fatol=None, frtol=0, maxiter=_iter, callback=None):
     """Find the root of an elementwise function using Chandrupatla's algorithm.
 
-    This function allows for `a`, `b`, amd the output of `f` to be of any
-    broadcastable shapes. For each element of the output of `f`, `chandrupatla`
-    seeks the scalar root that makes the element 0.
+    For each element of the output of `f`, `chandrupatla` seeks the scalar root
+    that makes the element 0. This function allows for `a`, `b`, and the output
+    of `f` to be of any broadcastable shapes.
 
     Parameters
     ----------
@@ -1421,12 +1421,9 @@ def _chandrupatla(func, a, b, *, args=(), xrtol=_xtol, xatol=_rtol,
         being passed into `func`.
     args : tuple, optional
         Additional positional arguments to be passed to `func`.
-    xatol : float, optional
-        Tolerance (absolute) for termination. Termination occurs if
-        `2*xatol*np.abs(r) + xrtol / np.abs(b - c) > 0.5`.
-    xrtol : float, optional
-        Tolerance (relative) for termination. Termination occurs if
-        `2*xatol*np.abs(r) + xrtol / np.abs(b - c) > 0.5`.
+    xatol, xrtol, fatol, frtol : float, optional
+        Absolute and relative x- and f- tolerances for termination. See Notes
+        for details.
     maxiter : int, optional
         The maximum number of iterations of the algorithm to perform.
     callback : callable, optional
@@ -1471,6 +1468,17 @@ def _chandrupatla(func, a, b, *, args=(), xrtol=_xtol, xatol=_rtol,
     -----
     Implemented based on Chandrupatla's original paper [1]_.
 
+    If ``xl`` and ``xr`` are the left and right ends of the bracket,
+    ``xmin = xl if abs(func(xl)) <= abs(func(xr)) else xr``,
+    and ``fmin0 = min(func(a), func(b))``, then the algorithm is considered to
+    have converged when ``abs(xr - xl) < xatol + abs(xmin) * xrtol`` or
+    ``fun(xmin) <= fatol + abs(fmin0) * frtol``. This is equivalent to the
+    termination condition described in [1]_ with ``xrtol = 4e-10``,
+    ``xatol = 1e-5``, and ``fatol = frtol = 0``. The default values are
+    ``xatol = 2e-12``, ``xrtol = 4 * np.finfo(float).eps``, ``frtol = 0``,
+    and ``fatol`` is the smallest normal number of the ``dtype`` returned
+    by ``func``.
+
     References
     ----------
 
@@ -1501,16 +1509,23 @@ def _chandrupatla(func, a, b, *, args=(), xrtol=_xtol, xatol=_rtol,
 
     """
 
-    def check_convergence(x1, f1, x2, f2, flag, eps_r=1e-10, eps_a=1e-5):
+    def check_convergence(x1, f1, x2, f2, flag, xatol, xrtol, fatol, frtol):
         # See Section 4
         xmin, fmin = x2.copy(), f2.copy()
         imin = np.abs(f1) < np.abs(f2)
         xmin[imin], fmin[imin] = x1[imin], f1[imin]
 
[…]
