### 2023-05-10 9e0a520
MAINT: optimize._chandrupatla: improve input validation, special cases
