+++ scipy/optimize/_zeros_py.py
@@ -1509,64 +1509,24 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
 
     """
 
-    def check_convergence(x1, f1, x2, f2, flag, xatol, xrtol, fatol, frtol):
-        # See Section 4
-        xmin, fmin = x2.copy(), f2.copy()
-        imin = np.abs(f1) < np.abs(f2)
-        xmin[imin], fmin[imin] = x1[imin], f1[imin]
+    res = _chandrupatla_iv(func, a, b, args, xatol, xrtol,
+                           fatol, frtol, maxiter, callback)
+    func, a, b, args, xatol, xrtol, fatol, frtol, maxiter, callback = res
 
-        # This is the criterion used in bisect. Chandrupatla's criterion
-        # is equivalent to abs(x2-x1) < 4*abs(xmin)*xrtol + xatol.
-        dx = abs(x2 - x1)
-        tol = abs(xmin)*xrtol + xatol
-        itol =  dx < tol
-        # Tolerance on function value. Note that `frtol` has been redefined as
-        # `frtol = frtol * np.minimum(f1, f2)`, where `f1` and `f2` are the
-        # function evaluated at the original ends of the bracket.
-        itol |= np.abs(fmin) <= fatol + frtol
-        if np.all(itol):
-            pass
-        flag[itol] = _ECONVERGED
-
-        j = np.sign(f1) == np.sign(f2)
-        xmin[j], fmin[j], itol[j], flag[j] = np.nan, np.nan, True, _ESIGNERR
-
-        j = (np.isfinite(x1) & np.isfinite(x2)
-             & np.isfinite(f1) & np.isfinite(f2))
-        itol[~j], flag[~j] = True, _EVALUEERR
-
-        tl = 0.5 * tol / dx
-        return xmin, fmin, np.all(itol), flag, tl
-
-    x1, x2 = np.broadcast_arrays(a, b)  # broadcast and rename
-    f1, f2 = func(x1[()], *args), func(x2[()], *args)
-
-    # need to broadcast again; need different brackets for each output of func
-    x1, x2, f1, f2 = np.broadcast_arrays(x1, x2, f1, f2)
-    # and they need to be writable floats
-    xt = np.result_type(x1.dtype, x2.dtype, np.float16)  # at least float16
-    ft = np.result_type(f1.dtype, f2.dtype, np.float16)
-    x1, x2, f1, f2 = x1.astype(xt), x2.astype(xt), f1.astype(ft), f2.astype(ft)
-
-    output_shape = x1.shape
-    x1, x2, f1, f2 = np.atleast_1d(x1, x2, f1, f2)
+    # Initialization
+    x1, f1, x2, f2, shape, dtype = _initialize_xf(func, a, b, args)
     flag = np.full_like(x1, _EINPROGRESS, dtype=int)  # in progress
     i, f_evals = 0, 2  # two function evaluations performed above
     cb_terminate = False
-    fatol = fatol or np.finfo(ft).smallest_normal
+    fatol = fatol or np.finfo(dtype).smallest_normal
     frtol = frtol * np.minimum(np.abs(f1), np.abs(f2))
+    tols = dict(xatol=xatol, xrtol=xrtol, fatol=fatol, frtol=frtol)
 
-    if not np.issubdtype(np.result_type(x1, x2, f1, f2), np.number):
-        message = "Bracket and function output must be numeric."
-        raise ValueError(message)
-
-    xmin, fmin, converged, flag, tl = (
-        check_convergence(x1, f1, x2, f2, flag,
-                          xatol=xatol, xrtol=xrtol, fatol=fatol, frtol=frtol))
-
+    xmin, fmin, converged, flag, tl = _check_termination(x1, f1, x2, f2, flag,
[…]
