### 2023-05-10 9f976cd
TST: optimize._chandrupatla: add and improve tests, comments
