+++ scipy/optimize/tests/test_zeros.py
@@ -213,17 +213,21 @@ class TestChandrupatla(TestScalarRootFinders):
 
     @pytest.mark.parametrize('p', [0.6, np.linspace(-0.05, 1.05, 10)])
     def test_basic(self, p):
+        # Invert distribution CDF and compare against distrtibution `ppf`
         dist = stats.norm()
-        with np.errstate(invalid='ignore', divide='ignore'):
+        with np.errstate(invalid='ignore'):
             res = zeros._chandrupatla(self.f, -5, 5, args=(dist, p))
         ref = dist.ppf(p)
         np.testing.assert_allclose(res.root, ref)
         assert res.root.shape == ref.shape
 
-    @pytest.mark.parametrize('shape', [(12,), (3, 4), (3, 2, 2)])
+    @pytest.mark.parametrize('shape', [tuple(), (12,), (3, 4), (3, 2, 2)])
     def test_vectorization(self, shape):
-        p = np.linspace(-0.05, 1.05, 12).reshape(shape)
+        # Test for correct functionality, output shapes, and dtypes for various
+        # input shapes.
+        p = np.linspace(-0.05, 1.05, 12).reshape(shape) if shape else 0.6
         dist = stats.norm()
+        args = (dist, p)
 
         @np.vectorize
         def chandrupatla_single(p):
@@ -234,13 +238,18 @@ class TestChandrupatla(TestScalarRootFinders):
             return self.f(*args, **kwargs)
         f.f_evals = 0
 
-        with np.errstate(invalid='ignore', divide='ignore'):
-            res = zeros._chandrupatla(f, -5, 5, args=(dist, p))
+        with np.errstate(invalid='ignore'):
+            res = zeros._chandrupatla(f, -5, 5, args=args)
             refs = chandrupatla_single(p).ravel()
+
+        ref_root = [ref.root for ref in refs]
+        assert_allclose(res.root.ravel(), ref_root)
+        assert_equal(res.root.shape, shape)
 
         ref_fun = [ref.fun for ref in refs]
         assert_allclose(res.fun.ravel(), ref_fun, atol=1e-9)
         assert_equal(res.fun.shape, shape)
+        assert_equal(res.fun, self.f(res.root, *args))
 
         ref_converged = [ref.converged for ref in refs]
         assert_equal(res.converged.ravel(), ref_converged)
@@ -264,24 +273,24 @@ class TestChandrupatla(TestScalarRootFinders):
         assert_equal(res.iterations.shape, res.fun.shape)
         assert np.issubdtype(res.iterations.dtype, np.integer)
 
-        ref_root = [ref.root for ref in refs]
-        assert_allclose(res.root.ravel(), ref_root)
-        assert_equal(res.root.shape, shape)
+        # test upper and lower bracket ends and function values
 
     def test_flags(self):
+        # Test cases that should produce different status flags; show that all
+        # can be produced simultaneously.
         def f(x):
             return [x[0] - 2.5, x[1] - 10, (x[2]-0.1)**3, np.nan]
 
-        with np.errstate(invalid='ignore', divide='ignore'):
-            res = zeros._chandrupatla(f, [0] * 4, [np.pi] * 4, maxiter=2)
+        res = zeros._chandrupatla(f, [0] * 4, [np.pi] * 4, maxiter=2)
 
         ref_flags = np.array([zeros._ECONVERGED, zeros._ESIGNERR,
                               zeros._ECONVERR, zeros._EVALUEERR])
         assert_equal(res.flag, ref_flags)
 
-    def test_termination(self):
-        # make this work with size > 1
-        # make sure upper_bracket and lower_bracket are ordered
+    def test_convergence(self):
[…]
