### 2023-05-13 0e9a3f3
TST: optimize._chandrupatla: additional tests
