### 2023-05-13 8e90b3e
MAINT: optimize._chandrupatla: rename bracket endpoints and function values
