### 2023-05-13 fd32788
MAINT: opimize._chandrupatla: convert RootResult to OptimizeResult
