+++ scipy/optimize/_zeros_py.py
@@ -1442,21 +1442,21 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
         will be scalars; however, if `func` returns an array, the outputs will
         be arrays of the same shape.
 
-        root : float
+        x : float
             The root of the function.
-        function_calls : int
+        nfev : int
             The number of times the function was called to find the root.
-        iterations : int
+        nit : int
             The number of iterations of Chandrupatla's algorithm performed.
-        flag : int
+        status : int
             An integer representing the exit status of the algorithm.
             ``0`` : The algorithm converged to the specified tolerances.
             ``-1`` : The algorithm encountered an invalid bracket.
             ``-2`` : The maximum number of iterations was reached.
             ``-3`` : A non-finite value was encountered.
             ``1`` : The algorithm is proceeding normally (in `callback` only).
-        converged : bool
-            ``True`` when the algorithm terminated successfully (flag ``0``).
+        success : bool
+            ``True`` when the algorithm terminated successfully (status ``0``).
         fun : float
             The function value evaluated at the root.
         xl, xr : float
@@ -1499,12 +1499,12 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
     ...     return x**3 - 2*x - c
     >>> c = 5
     >>> res = optimize._zeros_py._chandrupatla(f, 0, 3, args=(c,))
-    >>> res.root
+    >>> res.x
     2.0945514818937463
 
     >>> c = [3, 4, 5]
     >>> res = optimize._zeros_py._chandrupatla(f, 0, 3, args=(c,))
-    >>> res.root
+    >>> res.x
     array([1.8932892 , 2.        , 2.09455148])
 
     """
@@ -1515,19 +1515,24 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
 
     # Initialization
     x1, f1, x2, f2, shape, dtype = _initialize_xf(func, a, b, args)
-    flag = np.full_like(x1, _EINPROGRESS, dtype=int)  # in progress
+    status = np.full_like(x1, _EINPROGRESS, dtype=int)  # in progress
     i, f_evals = 0, 2  # two function evaluations performed above
     cb_terminate = False
+    nfev = np.full_like(status, f_evals)[()]
+    nit = np.full_like(status, i)[()]
     fatol = fatol or np.finfo(dtype).smallest_normal
     frtol = frtol * np.minimum(np.abs(f1), np.abs(f2))
     tols = dict(xatol=xatol, xrtol=xrtol, fatol=fatol, frtol=frtol)
 
-    xmin, fmin, converged, flag, tl = _check_termination(x1, f1, x2, f2, flag,
-                                                         **tols)
+    xmin, fmin, converged, status, tl = _check_termination(x1, f1, x2, f2,
+                                                           status, **tols)
+
+    res = OptimizeResult(x=xmin, fun=fmin, nit=nit, nfev=nfev, status=status,
+                         success=(status==0), xl=x1, fl=f1, xr=x2, fr=f2)
+
     if callback is not None:
-        res = _prepare_result(xmin, fmin, x1, f1, x2, f2,
-                              flag, f_evals, i, shape)
-        if _call_callback_maybe_halt(callback, res):
+        temp_res = _prepare_result(shape, res)
[…]
