+++ scipy/optimize/tests/test_zeros.py
@@ -218,8 +218,8 @@ class TestChandrupatla(TestScalarRootFinders):
         with np.errstate(invalid='ignore'):
             res = zeros._chandrupatla(self.f, -5, 5, args=(dist, p))
         ref = dist.ppf(p)
-        np.testing.assert_allclose(res.root, ref)
-        assert res.root.shape == ref.shape
+        np.testing.assert_allclose(res.x, ref)
+        assert res.x.shape == ref.shape
 
     @pytest.mark.parametrize('shape', [tuple(), (12,), (3, 4), (3, 2, 2)])
     def test_vectorization(self, shape):
@@ -242,36 +242,36 @@ class TestChandrupatla(TestScalarRootFinders):
             res = zeros._chandrupatla(f, -5, 5, args=args)
             refs = chandrupatla_single(p).ravel()
 
-        ref_root = [ref.root for ref in refs]
-        assert_allclose(res.root.ravel(), ref_root)
-        assert_equal(res.root.shape, shape)
+        ref_x = [ref.x for ref in refs]
+        assert_allclose(res.x.ravel(), ref_x)
+        assert_equal(res.x.shape, shape)
 
         ref_fun = [ref.fun for ref in refs]
         assert_allclose(res.fun.ravel(), ref_fun, atol=1e-9)
         assert_equal(res.fun.shape, shape)
-        assert_equal(res.fun, self.f(res.root, *args))
+        assert_equal(res.fun, self.f(res.x, *args))
 
-        ref_converged = [ref.converged for ref in refs]
-        assert_equal(res.converged.ravel(), ref_converged)
-        assert_equal(res.converged.shape, shape)
-        assert np.issubdtype(res.converged.dtype, np.bool_)
+        ref_success = [ref.success for ref in refs]
+        assert_equal(res.success.ravel(), ref_success)
+        assert_equal(res.success.shape, shape)
+        assert np.issubdtype(res.success.dtype, np.bool_)
 
-        ref_flag = [ref.flag for ref in refs]
-        assert_equal(res.flag.ravel(), ref_flag)
-        assert_equal(res.flag.shape, shape)
-        assert np.issubdtype(res.flag.dtype, np.integer)
+        ref_flag = [ref.status for ref in refs]
+        assert_equal(res.status.ravel(), ref_flag)
+        assert_equal(res.status.shape, shape)
+        assert np.issubdtype(res.status.dtype, np.integer)
 
-        ref_function_calls = [ref.function_calls for ref in refs]
-        assert_equal(res.function_calls, max(ref_function_calls))
-        assert_equal(res.function_calls, f.f_evals)
-        assert_equal(res.function_calls.shape, res.fun.shape)
-        assert np.issubdtype(res.function_calls.dtype, np.integer)
+        ref_nfev = [ref.nfev for ref in refs]
+        assert_equal(res.nfev, max(ref_nfev))
+        assert_equal(res.nfev, f.f_evals)
+        assert_equal(res.nfev.shape, res.fun.shape)
+        assert np.issubdtype(res.nfev.dtype, np.integer)
 
-        ref_iterations = [ref.iterations for ref in refs]
-        assert_equal(res.iterations, max(ref_iterations))
-        assert_equal(res.iterations, f.f_evals-2)
-        assert_equal(res.iterations.shape, res.fun.shape)
-        assert np.issubdtype(res.iterations.dtype, np.integer)
+        ref_nit = [ref.nit for ref in refs]
+        assert_equal(res.nit, max(ref_nit))
+        assert_equal(res.nit, f.f_evals-2)
+        assert_equal(res.nit.shape, res.fun.shape)
+        assert np.issubdtype(res.nit.dtype, np.integer)
 
         # test upper and lower bracket ends and function values
 
[…]
