### 2023-05-13 25f3594
MAINT: optimize._chandrupatla: always work with 1D arrays
