### 2023-05-13 aab8570
MAINT: optimize._chandrupatla: avoid unnecessary calculation
