+++ scipy/optimize/_zeros_py.py
@@ -1514,7 +1514,7 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
     func, a, b, args, xatol, xrtol, fatol, frtol, maxiter, callback = res
 
     # Initialization
-    x1, f1, x2, f2, shape, dtype = _initialize_xf(func, a, b, args)
+    x1, f1, x2, f2, active, shape, dtype = _initialize_xf(func, a, b, args)
     status = np.full_like(x1, _EINPROGRESS, dtype=int)  # in progress
     i, f_evals = 0, 2  # two function evaluations performed above
     cb_terminate = False
@@ -1524,11 +1524,13 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
     frtol = frtol * np.minimum(np.abs(f1), np.abs(f2))
     tols = dict(xatol=xatol, xrtol=xrtol, fatol=fatol, frtol=frtol)
 
-    xmin, fmin, converged, status, tl = _check_termination(x1, f1, x2, f2,
-                                                           status, **tols)
+    res = OptimizeResult(x=x1.copy(), fun=x2.copy(), nit=nit, nfev=nfev,
+                         status=status.copy(), success=(status==0),
+                         xl=x1.copy(), fl=f1.copy(), xr=x2.copy(),
+                         fr=f2.copy())
 
-    res = OptimizeResult(x=xmin, fun=fmin, nit=nit, nfev=nfev, status=status,
-                         success=(status==0), xl=x1, fl=f1, xr=x2, fr=f2)
+    temp = _check_termination(x1, f1, x2, f2, x1, f1, res, active, status, f_evals, i, tols)
+    xmin, fmin, x1, f1, x2, f2, x3, f3, converged, active, status, tl = temp
 
     if callback is not None:
         temp_res = _prepare_result(shape, res)
@@ -1539,7 +1541,10 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
     while np.all(i < maxiter) and not converged and not cb_terminate:
         # Flowchart 1
         x = x1 + t * (x2 - x1)
-        f = np.asarray(func(x.reshape(shape), *args), dtype=dtype).ravel()
+        X = res.x.copy()
+        X[active] = x
+        f = np.asarray(func(X.reshape(shape), *args), dtype=dtype).ravel()
+        f = f[active]
         f_evals += 1
 
         # # Flowchart 2 (flowchart is reversed; see code)
@@ -1552,14 +1557,16 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
 
         # Flowchart 3
         i += 1
-        xmin, fmin, converged, status, tl = _check_termination(x1, f1, x2, f2,
-                                                               status, **tols)
-        res = _update_result(xmin, fmin, x1, f1, x2, f2, status, f_evals, i, res)
+        temp = _check_termination(x1, f1, x2, f2, x3, f3, res, active, status, f_evals, i, tols)
+        xmin, fmin, x1, f1, x2, f2, x3, f3, converged, active, status, tl = temp
+
         if callback is not None:
             temp_res = _prepare_result(shape, res)
             if _call_callback_maybe_halt(callback, temp_res) or converged:
                 cb_terminate = True
                 break
+        if converged:
+            break
 
         # Flowchart 4
         # Equation 1
[…]
