### 2023-05-13 e663d82
MAINT: optimize._chandrupatla: fix iteration count in callback
