+++ scipy/optimize/_zeros_py.py
@@ -1516,36 +1516,36 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
     # Initialization
     x1, f1, x2, f2, active, shape, dtype = _initialize_xf(func, a, b, args)
     status = np.full_like(x1, _EINPROGRESS, dtype=int)  # in progress
-    i, f_evals = 0, 2  # two function evaluations performed above
+    nit, nfev = 0, 2  # two function evaluations performed above
     cb_terminate = False
-    nfev = np.full_like(status, f_evals)[()]
-    nit = np.full_like(status, i)[()]
     fatol = fatol or np.finfo(dtype).smallest_normal
     frtol = frtol * np.minimum(np.abs(f1), np.abs(f2))
     tols = dict(xatol=xatol, xrtol=xrtol, fatol=fatol, frtol=frtol)
 
-    res = OptimizeResult(x=x1.copy(), fun=x2.copy(), nit=nit, nfev=nfev,
+    res = OptimizeResult(x=x1.copy(), fun=x2.copy(),
+                         nit=np.full_like(status, nit)[()],
+                         nfev=np.full_like(status, nfev)[()],
                          status=status.copy(), success=(status==0),
                          xl=x1.copy(), fl=f1.copy(), xr=x2.copy(),
                          fr=f2.copy())
 
-    temp = _check_termination(x1, f1, x2, f2, x1, f1, res, active, status, f_evals, i, tols)
+    temp = _check_termination(x1, f1, x2, f2, x1, f1, res, active, status, nfev, nit, tols)
     xmin, fmin, x1, f1, x2, f2, x3, f3, converged, active, status, tl = temp
 
     if callback is not None:
-        temp_res = _prepare_result(shape, res)
+        temp_res = _prepare_result(shape, res, active, nit, nfev)
         if _call_callback_maybe_halt(callback, temp_res):
             cb_terminate == True
 
     t = 0.5
-    while np.all(i < maxiter) and not converged and not cb_terminate:
+    while nit < maxiter and not converged and not cb_terminate:
         # Flowchart 1
         x = x1 + t * (x2 - x1)
         X = res.x.copy()
         X[active] = x
         f = np.asarray(func(X.reshape(shape), *args), dtype=dtype).ravel()
         f = f[active]
-        f_evals += 1
+        nfev += 1
 
         # # Flowchart 2 (flowchart is reversed; see code)
         x3, f3 = x2.copy(), f2.copy()
@@ -1556,12 +1556,12 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
         x1, f1 = x, f
 
         # Flowchart 3
-        i += 1
-        temp = _check_termination(x1, f1, x2, f2, x3, f3, res, active, status, f_evals, i, tols)
+        nit += 1
+        temp = _check_termination(x1, f1, x2, f2, x3, f3, res, active, status, nfev, nit, tols)
         xmin, fmin, x1, f1, x2, f2, x3, f3, converged, active, status, tl = temp
 
         if callback is not None:
-            temp_res = _prepare_result(shape, res)
+            temp_res = _prepare_result(shape, res, active, nit, nfev)
             if _call_callback_maybe_halt(callback, temp_res) or converged:
                 cb_terminate = True
                 break
@@ -1583,10 +1583,8 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
[…]
