### 2023-05-13 b28577f
MAINT: optimize._chandrupatla: cleanup / refactor
