+++ scipy/optimize/_zeros_py.py
@@ -1416,9 +1416,8 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
          `_chandrupatla` seeks an array (or scalar) ``x`` such that each
          corresponding element in the output array (or scalar) is zero.
     a, b : array_like
-        The lower and upper bounds of the root of the function. Follows normal
-        broadcasting rules and will be broadcasted against one another before
-        being passed into `func`.
+        The lower and upper bounds of the root of the function. Must be
+        broadcastable with one another.
     args : tuple, optional
         Additional positional arguments to be passed to `func`.
     xatol, xrtol, fatol, frtol : float, optional
@@ -1427,23 +1426,24 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
     maxiter : int, optional
         The maximum number of iterations of the algorithm to perform.
     callback : callable, optional
-        An optional user-supplied function to be called after each iteration.
-        Called as ``callback(res)``, where ``res`` is a ``RootResult``
-        like that returned by `_chandrupatla` but containing the current
-        iterate's values of all variables. If the raises a ``StopIteration``,
-        the algorithm will terminate immediately and `_chandrupatla` will
-        return a result.
+        An optional user-supplied function to be called before the first
+        iteration and after each iteration.
+        Called as ``callback(res)``, where ``res`` is an ``OptimizeResult``
+        similar to that returned by `_chandrupatla` (but containing the current
+        iterate's values of all variables). If `callback` raises a
+        ``StopIteration``, the algorithm will terminate immediately and
+        `_chandrupatla` will return a result.
 
     Returns
     -------
-    res : RootResult
-        An instance of a subclass of `scipy.optimize.OptimizeResult` with the
-        following attributes. The descriptions are written as though the values
-        will be scalars; however, if `func` returns an array, the outputs will
-        be arrays of the same shape.
+    res : OptimizeResult
+        An instance of `scipy.optimize.OptimizeResult` with the following
+        attributes. The descriptions are written as though the values will be
+        scalars; however, if `func` returns an array, the outputs will be
+        arrays of the same shape.
 
         x : float
-            The root of the function.
+            The root of the function, if the algorithm terminated successfully.
         nfev : int
             The number of times the function was called to find the root.
         nit : int
@@ -1454,11 +1454,12 @@ def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
             ``-1`` : The algorithm encountered an invalid bracket.
             ``-2`` : The maximum number of iterations was reached.
             ``-3`` : A non-finite value was encountered.
+            ``-4`` : Iteration was terminated by `callback`.
             ``1`` : The algorithm is proceeding normally (in `callback` only).
         success : bool
             ``True`` when the algorithm terminated successfully (status ``0``).
[…]
