### 2023-05-16 d366d28
MAINT: optimize._bracket_root: begin to vectorize
