+++ paper.bib
@@ -1,4 +1,4 @@
-@article{fisher1949design,
+@book{fisher1949design,
   title={The design of experiments},
   author={Fisher, Ronald A},
   year={1949},
@@ -23,7 +23,7 @@
 }
 
 @book{robert1999monte,
-  title={Monte Carlo statistical methods},
+  title={Monte {C}arlo statistical methods},
   author={Robert, Christian P and Casella, George and Casella, George},
   volume={2},
   year={1999},
@@ -38,7 +38,7 @@
 }
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -90,7 +90,7 @@
 
 @misc{ye2020,
   author = {Leihua Ye},
-  title = {Bootstrap and Statistical Inference in Python},
+  title = {Bootstrap and Statistical Inference in {P}ython},
   year = {2020},
   month = {12},
   publisher = {Medium},
@@ -111,7 +111,7 @@
 
 @misc{omar2021,
   author = {Hamid Omar},
-  title = {A Guide on How to Simulate and Visualize Hypothesis Tests Using Python Code},
+  title = {A Guide on How to Simulate and Visualize Hypothesis Tests Using {P}ython Code},
   year = {2021},
   month = {06},
   publisher = {Medium},
