### 2023-06-12 #5092 @jbytecode
@mdhaber - One more thing, could you please add the ISBN code to the Fisher(1949) book in your bibtex file?
