### 2023-06-13 797a7d7
Update paper.bib
Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>
