+++ paper.bib
@@ -1,7 +1,8 @@
 @book{fisher1949design,
   title={The design of experiments},
   author={Fisher, Ronald A},
-  year={1949},
+  year={1951},
+  isbn={9780050008683},
   publisher={Oliver \& Boyd}
 }
 
