### 2023-06-13 #5092 @jbytecode
@check references
