### 2023-06-13 #5092 @mdhaber
Sure, I'll work on that. Should I reformat these citations?
![image](https://github.com/openjournals/joss-reviews/assets/6570539/d88cffce-7011-43e8-8134-ba295813feb0)

There probably should not be double parentheses. Do you remember offhand the syntax for removing those?
