### 2023-06-13 #5092 @jbytecode
@mdhaber - sure, I think you don't need an e.g. here and just (Omar, 2021; Saenger, 2019; Ye, 2020) is enough.
