+++ paper.md
@@ -43,7 +43,7 @@ Monte Carlo hypothesis tests, permutation tests, and the bootstrap are general p
 
 Although there are a variety of specialized statistical procedures for answering such questions (e.g., normality tests, t-tests, standard error of the mean), most computer implementations provide answers that are only guaranteed to be accurate when assumptions that simplify the mathematics are met (e.g., the sample size is large, there are no ties in the data). At the expense of additional computation, the `scipy.stats` functions `monte_carlo_test`, `permutation_test`, and `bootstrap` provide accurate answers to questions 1-3 even when the simplifying assumptions of specialized procedures are not met. Furthermore, these functions can be used to answer questions 1-3 even when specialized procedures have not been *developed*, giving users tremendous flexibility in the analysis that they can perform.
 
-Before the release of SciPy 1.9.0, the need for these procedures was partially met in the scientific Python ecosystem by tutorials (e.g., [@ye2020; @saenger2019; @omar2021]) and niche packages (e.g., [@resample2020; @permute2021]), but the functions in SciPy have several advantages:
+Before the release of SciPy 1.9.0, the need for these procedures was partially met in the scientific Python ecosystem by tutorials [@ye2020; @saenger2019; @omar2021] and niche packages [@resample2020; @permute2021], but the functions in SciPy have several advantages:
 
 - Prevalence: SciPy is one of the most frequently downloaded scientific Python packages. If a Python user finds the need for these statistical methods, chances are that they already have SciPy, eliminating the need to find and install a new package.
 - Speed: the functions take advantage of vectorized user code, avoiding slow Python loops.
