### 2023-06-13 #5092 @mdhaber
Here is the release: https://github.com/mdhaber/scipy/releases/tag/v1.10.0-joss-article
