### 2023-06-13 #5092 @mdhaber
Good to know. But I went ahead and recreated it. Same URL.
https://github.com/mdhaber/scipy/releases/tag/v1.10.0-joss-article
