### 2023-06-13 #5092 @jbytecode
@mdhaber - In the software repo, it seems the license is `BSD-3-Clause license` whereas in Zenodo archive it is  `Creative Commons Attribution 4.0 International`. Could you please correct that and report the changes here
