### 2023-06-13 #5092 @mdhaber
Thanks. I must have been fixing that as you writing this. I think it fixed the license at the original DOI.
![image](https://github.com/openjournals/joss-reviews/assets/6570539/97495d19-0234-423d-9467-1ec36503c414)
